+++ src/app/lib/widgets/observation-link-period/observation-link-period.component.ts
@@ -59,6 +59,7 @@ export class ObservationLinkPeriodComponent extends StringComponent implements O
    * Read extension and initialize properties.
    */
   ngOnInit() {
+    super.ngOnInit();
     this.setOlp();
     this.extensionsService.extensionsObservable.subscribe(() => {
       this.setOlp();